+++ cucumber/policy/features/step_definitions/error_steps.rb
@@ -5,3 +5,13 @@ end
 Then(/^the error message is "([^"]*)"$/) do |message|
   expect(@error['message']).to eq(message)
 end
+
+Then(/^there is an error$/) do
+  json = json_result
+  expect(json).to have_key('error')
+  @error = json['error']
+end
+
+Then(/^there is no error$/) do
+  expect(json_result).not_to have_key('error')
+end