+++ cucumber/policy/features/support/world.rb
@@ -15,19 +15,21 @@ end
 
 module PossumWorld
   include FullId
-  
+
   attr_reader :result
 
-  def invoke status: :ok, &block
+  # invoke accepts an optional HTTP status code as input
+  # and checks that the result matches that code
+  def invoke status: nil, &block
     begin
       @result = yield
-      raise "Expected invocation to be denied" unless status == :ok
+      raise "Expected invocation to be denied" if status && status != 200
       @result.tap do |result|
         puts result if @echo
       end
     rescue RestClient::Exception => e
-      @exception = e
-      raise e unless status == e.http_code
+      expect(e.http_code).to eq(status) if status
+      @result = e.response.body
     end
   end
 
@@ -42,7 +44,7 @@ module PossumWorld
   def extend_root_policy policy
     conjur_api.load_policy "root", policy, method: Conjur::API::POLICY_METHOD_POST
   end
-  
+
   def load_policy id, policy
     conjur_api.load_policy id, policy, method: Conjur::API::POLICY_METHOD_PUT
   end
@@ -54,20 +56,29 @@ module PossumWorld
   def extend_policy id, policy
     conjur_api.load_policy id, policy, method: Conjur::API::POLICY_METHOD_POST
   end
-  
+
   def make_full_id *tokens
     super tokens.join(":")
   end
-  
+
   def conjur_api
     login_as_role 'admin', admin_api_key unless @conjur_api
     @conjur_api
   end
 
+  def json_result
+    case @result
+    when String
+      JSON.parse(@result)
+    when Conjur::PolicyLoadResult
+      JSON.parse(@result.to_json)
+    end
+  end
+
   def admin_api_key
     @admin_api_key ||= Conjur::API.login 'admin', admin_password
   end
-  
+
   def admin_password
     ENV['CONJUR_AUTHN_PASSWORD'] || 'admin'
   end