+++ bin/conjur-cli.rb
@@ -68,7 +68,16 @@ command :server do |c|
       system "rake policy:load[#{account},#{file_name}]" or exit $?.exitstatus
     end
 
-    exec "rails server -p #{options[:port]} -b #{options[:'bind-address']}"
+
+    webserver_pid = Process.fork do
+      exec "rails server -p #{options[:port]} -b #{options[:'bind-address']}"
+    end
+    authn_local_pid = Process.fork do
+      exec "rake authn_local:run"
+    end
+
+    Process.wait webserver_pid
+    Process.wait authn_local_pid
   end
 end