+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Build scripts now look at git tags to determine version and tags to use.
 
 ### Fixed
-- Foreign key constraint violation is handled for violations of `role_memberships_role_id_fkey` or `role_memberships_member_id_fkey`
+- When a policy is loaded which references a non-existant object, that error is now reported as a JSON-formatted 404 error rather than an ugly 500 error.
 
 ## 0.1.0 - 2017-12-04